Fix issue #220, detection of zfsutils package in Ubuntu
This fixes issue #220. Previously, the script automatically added
`zfsutils` as dependecy. In Ubuntu 16.04 the package is called
zfsutils-linux. This commit runs `lsb_release -r` to get the release
number and then checks if it is running on 16.04 or greater to set the
correct zfsutils dependency.

Note, this commit was only tested on an Ubuntu machine. It may need
editing for Debian or other distros.