Option to clean source snapshots even if a destination failed
Useful if one or more of the destinations is regularly not available but
you don't want source snapshots to be forever building up. It's an
option because it's slightly more dangerous than simply not destroying
snapshots in the event of a failure.

jim> Updated for handling also the protection of newest common snapshots
during recursive cleanup of the policy's root dataset - a mode which was
added in master later than the original PRed code branched off.